+++ src/p3.cpp
@@ -17,7 +17,7 @@ void P3::write_image() {
 
     for (int y = 0; y < m_img_height; y++) {
         for (int x = 0; x < m_img_width; x++) {
-            Vector3<float> vec = m_buffer[x + m_img_width * y];
+            Vector3<double> vec = m_buffer[x + m_img_width * y];
             m_outfile << static_cast<int>(255.999 * vec[0]) << " "
                       << static_cast<int>(255.999 * vec[1]) << " "
                       << static_cast<int>(255.999 * vec[2]) << " ";
@@ -27,7 +27,7 @@ void P3::write_image() {
     }
 }
 
-void P3::set_img_buffer(std::vector<Vector3<float>> &buffer) {
+void P3::set_img_buffer(std::vector<Vector3<double>> &buffer) {
     m_buffer = std::move(buffer);
 }
