+++ src/ray.cpp
@@ -1,5 +1,5 @@
 #include "ray.h"
 
-Vector3<float> Ray::at(float scalar) const {
+Vector3<double> Ray::at(double scalar) const {
     return m_origin + scalar * (m_direction - m_origin);
 }
